+++ plugins/action/fabric_summary_info.py
@@ -29,8 +29,8 @@ argument_spec = dnac_argument_spec()
 # Add arguments specific for this module
 argument_spec.update(
     dict(
-        startTime=dict(type="int"),
-        endTime=dict(type="int"),
+        startTime=dict(type="float"),
+        endTime=dict(type="float"),
         siteHierarchy=dict(type="str"),
         siteHierarchyId=dict(type="str"),
         headers=dict(type="dict"),
