+++ plugins/action/field_notices_results_notices_info.py
@@ -30,7 +30,7 @@ argument_spec = dnac_argument_spec()
 argument_spec.update(
     dict(
         id=dict(type="str"),
-        deviceCount=dict(type="int"),
+        deviceCount=dict(type="float"),
         type=dict(type="str"),
         offset=dict(type="int"),
         limit=dict(type="int"),
